% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{annotationDictionary}
\alias{annotationDictionary}
\title{Create a snapshot of the Annotation Dictionary}
\usage{
annotationDictionary(commonKnowledge)
}
\arguments{
\item{commonKnowledge}{The commonKnowledge data frame containing current annotations via pullAnnotations().}
}
\value{
Nothing.  Creates character vectors containing the `categorical` annotations, the `truefalse` annotations, the union of these `fieldList`, and all columns in `summarizeList`.
}
\description{
Pulls sample data down from REDCap in order to generate example column lists for use in Shiny UI's.
}
\details{
Requires **admin** REDCap credentials to be set in the environment.
}
\author{
Amy Paguirigan
}
