% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copyAnnotations.R
\name{copyAnnotations}
\alias{copyAnnotations}
\title{Copy annotations from another record}
\usage{
copyAnnotations(copyFrom = NULL, copyTo = NULL,
  type = c("submission", "study"), overwriteExisting = FALSE)
}
\arguments{
\item{copyFrom}{The molecular_id of the record to copy data from}

\item{copyTo}{The molecular_id of the record to edit using copied data}

\item{type}{A character array of the types of annotations you want to copy: "submission", "study", "subject", "biospecimen", "assaymaterial", "genomics". Default: c("submission", "study")}

\item{overwriteExisting}{Default to false.  If data exists in the record to edit, do you want to overwrite existing data?}
}
\value{
Results of the copy attempt including the data of the new record as well as the REDCap response.
}
\description{
Copies the study, specimen and assay material metadata from another record in the Repository
}
\details{
Requires valid credentials to be set in the environment with setCreds() and to have write priviledges to the data
}
\author{
Amy Paguirigan
}
