% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cromwellfunctions.R
\name{cromwellCache}
\alias{cromwellCache}
\title{Pull Cromwell Call Caching Data}
\usage{
cromwellCache(workflow_id)
}
\arguments{
\item{workflow_id}{The workflow ID to return call caching metadata for.}
}
\value{
Returns a long form data frame of metadata on call caching in a workflow.
}
\description{
Gets info about call caching status for the calls of a workflow
}
\details{
Requires valid Cromwell URL to be set in the environment.
}
\examples{
## Set credentials from a file with the specified format, called `secrets.R` in path `~/myCreds/`.
setCreds(tokenSet = "file", path = "~/R/requiredCredentials.R")
## Request what jobs have been submitted to your Cromwell instance in the past 7 days.
recentJobs <- cromwellJobs(days = 7)
## Request workflow metadata for a specific job that was run in your Cromwell instance.
thisWorkflowID <- recentJobs$workflow_id[1]
cacheMeta <- cromwellCache(workflow_id = thisWorkflowID)
}
\author{
Amy Paguirigan
}
