% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cromwellfunctions.R
\name{cromwellCall}
\alias{cromwellCall}
\title{Pull Cromwell Workflow Calls}
\usage{
cromwellCall(workflow_id)
}
\arguments{
\item{workflow_id}{The workflow ID to return call metadata on.}
}
\value{
Returns a long form data frame of metadata on calls.
}
\description{
Retrieve and process call data for a workflow
}
\details{
Requires valid Cromwell URL to be set in the environment.
}
\author{
Amy Paguirigan
}
