% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cromwellfunctions.R
\name{cromwellFailures}
\alias{cromwellFailures}
\title{Pull Cromwell Call Failure Data}
\usage{
cromwellFailures(workflow_id)
}
\arguments{
\item{workflow_id}{The workflow ID to return call failure metadata for.}
}
\value{
Returns a long form data frame of metadata on failed calls in a workflow.
}
\description{
Gets info about failed jobs for a workflow
}
\details{
Requires valid cromwell URL to be set in the environment.
}
\author{
Amy Paguirigan
}
