% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cromwellfunctions.R
\name{cromwellJobs}
\alias{cromwellJobs}
\title{Pull Cromwell Jobs}
\usage{
cromwellJobs(days = 7)
}
\arguments{
\item{days}{The number of days of history to return.}
}
\value{
Returns a long form data frame of metadata on workflows.
}
\description{
Requests metadata about Cromwell Jobs
}
\details{
Requires valid cromwell URL to be set in the environment.
}
\author{
Amy Paguirigan
}
