% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cromwellfunctions.R
\name{cromwellWorkflow}
\alias{cromwellWorkflow}
\title{Pull Cromwell Workflow Metadata}
\usage{
cromwellWorkflow(workflow_id)
}
\arguments{
\item{workflow_id}{The workflow ID to return metadata for.}
}
\value{
Returns a long form data frame of metadata on workflows.
}
\description{
Retrieve and process all labels, submission and workflow level metadata for a workflow
}
\details{
Requires valid cromwell URL to be set in the environment.
}
\author{
Amy Paguirigan
}
