% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDictionary.R
\name{getDictionary}
\alias{getDictionary}
\title{Retrieve an example data dictionary from the TGR REDCap project}
\usage{
getDictionary()
}
\value{
An array of field names for the project specified.
}
\description{
Retrieve an example data dictionary from the TGR REDCap project
}
\details{
Requires REDCap credentials to be set in the environment.
}
\author{
Amy Paguirigan
}
