% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTGRid.R
\name{getTGRid}
\alias{getTGRid}
\title{Request next subsequent molecular id in the TGR}
\usage{
getTGRid(DAG = NULL)
}
\arguments{
\item{DAG}{Data access group to assign the new record to.}
}
\value{
An empty data frame with the new identifier and it's data access group, ready for additions and upload to REDCap.  Also initializes the record in REDCap.
}
\description{
Requests and initializes a new record in REDCap in the specified Data Access Group (DAG)
}
\details{
Requires valid admin credentials to be set in the environment with setCreds()
}
\author{
Amy Paguirigan
}
