% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{listS3RepoObjects}
\alias{listS3RepoObjects}
\title{Pull the list of objects and tags in the Repository overall}
\usage{
listS3RepoObjects(bucket = "fh-pi-paguirigan-a-genomicsrepo")
}
\arguments{
\item{bucket}{The name of the S3 bucket containing the data, or "repository" if the intention is to query the whole Repository.}
}
\value{
Returns a long form data frame of annotated objects in the S3 bucket.
}
\description{
Pulls Information from the results of s3tagcrawler for TGR that are in an S3 bucket, including the object list and their tags as well as size metadata.
}
\details{
Requires S3 credentials to be set in the environment by setCreds.
}
\author{
Amy Paguirigan
}
