% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{listS3RepoSummaries}
\alias{listS3RepoSummaries}
\title{Pull the summary of objects and tags in the Repository overall}
\usage{
listS3RepoSummaries(bucket = "fh-pi-paguirigan-a-genomicsrepo")
}
\arguments{
\item{bucket}{The name of the S3 bucket containing the data.}
}
\value{
Returns a long form data frame of annotated objects in the S3 bucket.
}
\description{
Pulls just the processed summary of object metadata for all buckets with data in the Repository.
}
\details{
Requires **admin** or **app** S3 credentials to be set in the environment by setCreds.
}
\author{
Amy Paguirigan
}
