% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepForUpload.R
\name{prepForUpload}
\alias{prepForUpload}
\title{Separate a data frame by TGR REDCap Project for upload}
\usage{
prepForUpload(fileStem, dataToSplit)
}
\arguments{
\item{fileStem}{A character vector that you want the resulting csv's to have as the beginning of the filenames.}

\item{dataToSplit}{A data frame WITHOUT rownames, containing the data you want to split and upload. Note:  For a valid dataset to be uploaded to REDCap, any relevant Repository identifier needs to be included in the data frame.}
}
\value{
Writes up to 3 csv's ready for upload into the individual REDCap Projects of the TGR to the working directory.
}
\description{
Separate a data frame by TGR REDCap Project for upload
}
\details{
Requires REDCap credentials to be set in the environment.  Does not allow for changes in DAG or marking if an instrument is complete.
}
\author{
Amy Paguirigan
}
