% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgapi.R
\name{process_response}
\alias{process_response}
\title{Process JSON response into tibble}
\usage{
process_response(json, selected_fields = NULL)
}
\arguments{
\item{json}{JSON response}

\item{selected_fields}{Fields to include in output}
}
\value{
Tibble with selected fields
}
\description{
Process JSON response into tibble
}
