% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullAnnotations.R
\name{pullAnnotations}
\alias{pullAnnotations}
\title{Pull Current TGR Annotation definitions}
\usage{
pullAnnotations()
}
\value{
A data frame containing the current TGR Annotations, Values and their Definitions
}
\description{
Pulls current data about annotations from GitHub for use in annotating molecular data sets in the Repository.
}
\details{
Nothing to see here.
}
\author{
Amy Paguirigan
}
