% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcapPull.R
\name{redcapPull}
\alias{redcapPull}
\title{Queries all TGR REDCap for metadata}
\usage{
redcapPull(DAG = c("paguirigana", "bergera"), domain = "all",
  harmonizedOnly = FALSE, evenEmptyCols = FALSE)
}
\arguments{
\item{DAG}{The name of the data access group for which to request data. Current DAG options include: 'paguirigana', 'aberger'.  A character vector can be supplied.}

\item{domain}{The domain of the data type to return.  Values include: `all`, `specimen`, `assay`, `molecular`}

\item{harmonizedOnly}{Whether you want only the data that is harmonized across all projects (TRUE) or a complete data set (FALSE).}

\item{evenEmptyCols}{Whether you want even the empty columns in your final data frame (TRUE) or if you just want columns where there is at least one value in the resulting dataset (FALSE).}
}
\value{
Returns a long form data frame containing the requested dataset.
}
\description{
Queries all TGR REDCap for metadata
}
\details{
Requires REDCap credentials to be set in the environment.
}
\author{
Amy Paguirigan
}
