% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3functions.R
\name{s3_copy_and_tag}
\alias{s3_copy_and_tag}
\title{(admin) Copy and tag objects between S3 buckets}
\usage{
s3_copy_and_tag(fromBucket, fromPrefix, toBucket, toPrefix, tagList)
}
\arguments{
\item{fromBucket}{A character vector containing the full name of the source bucket}

\item{fromPrefix}{A character vector containing the prefix of the object to copy in the source bucket}

\item{toBucket}{A character vector containing the full name of the destination bucket}

\item{toPrefix}{A character vector containing the prefix to give to the object in the destination bucket}

\item{tagList}{A named list of tags to apply to the destination object}
}
\value{
TBD
}
\description{
Copies and tags objects in an S3 bucket from another S3 bucket.
}
\details{
Requires valid admin S3 credentials to be set in the environment by setCreds and the likely the output from cromwellOutput()
}
\author{
Amy Paguirigan
}
