% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setCreds.R
\name{setCreds}
\alias{setCreds}
\title{Sets tgR environment variables}
\usage{
setCreds(tokenSet = "admin")
}
\arguments{
\item{tokenSet}{Which token set to apply, currently either "admin" or "app".}
}
\value{
Nothing, sets environment variables for the remaining functions to work.
}
\description{
Sets tgR environment variables
}
\details{
Requires the desired REDCap and S3 credentials to be set in the environment.
}
\author{
Amy Paguirigan
}
