% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3functions.R
\name{summarizeS3Objects}
\alias{summarizeS3Objects}
\title{Pull a summary of objects in S3}
\usage{
summarizeS3Objects(bucket, prefix = "tg")
}
\arguments{
\item{bucket}{A character vector containing the full names of the S3 bucket(s) containing the data to return.}

\item{prefix}{The prefix of the data being indexed by the TGR (default: "tg").}
}
\value{
Returns a data frame containing a summary of what objects are in the indicated S3 bucket(s).
}
\description{
Pulls summary information from the apptags prefix in S3 for a given S3 bucket(s).
}
\details{
Requires valid S3 credentials to be set in the environment by setCreds.
}
\author{
Amy Paguirigan
}
