% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgR.R
\docType{package}
\name{tgR}
\alias{tgR}
\title{tgR: A convenience package for the Translational Genomics Repository.}
\description{
This package provides helper functions needed to coordinate between various aspects of repository.
}
