% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgrAnnotate.R
\name{tgrAnnotate}
\alias{tgrAnnotate}
\title{Queries TGR REDCap for dataset annotations}
\usage{
tgrAnnotate(DAG, harmonizedOnly = FALSE, evenEmptyCols = FALSE)
}
\arguments{
\item{DAG}{(Optional) A character vector contining the name(s) of the dtgrata access group(s) for which to request data, or "all" (parameter applies to admin only).}

\item{harmonizedOnly}{Whether you want only the annotations that are harmonized across all projects (TRUE) or a complete data set (FALSE).}

\item{evenEmptyCols}{Whether you want even the empty fields in your final data frame (TRUE) or if you just want columns where there is at least one value in the resulting dataset (FALSE).}
}
\value{
Returns a long form data frame containing the requested dataset.
}
\description{
Queries TGR REDCap for dataset annotations
}
\details{
Requires REDCap credentials to be set in the environment.  Was redcapPULL
}
\author{
Amy Paguirigan
}
