% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgrDefinitions.R
\name{tgrDefinedAnnotations}
\alias{tgrDefinedAnnotations}
\title{Pull Current TGR Annotation and Value definitions}
\usage{
tgrDefinedAnnotations()
}
\value{
A data frame containing the current TGR Annotations, Values and their Definitions
}
\description{
Pulls current data about annotations from GitHub for use in annotating molecular data sets in the Repository.
}
\author{
Amy Paguirigan
}
