% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotationfunctions.R
\name{tgrDefinitions}
\alias{tgrDefinitions}
\title{Pull Current TGR Annotation and Value definitions}
\usage{
tgrDefinitions()
}
\value{
A data frame containing the current TGR Annotations, Values and their Definitions

Returns a data frame containing TGR metadata definitions.
}
\description{
Pulls current data about annotations from GitHub for use in annotating molecular data sets in the Repository.
}
\details{
Nothing to see here.
}
\author{
Amy Paguirigan
}
