% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uherotgr-package.R
\docType{package}
\name{uherotgr-package}
\alias{uherotgr}
\alias{uherotgr-package}
\title{uherotgr: Wrapper for Title Guarantee's API}
\description{
Helper functions to access and store data from Title Guarantee's API.
}
\author{
\strong{Maintainer}: Victoria Rhinebolt \email{vward@hawaii.edu}

}
\keyword{internal}
