% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotationfunctions.R
\name{undefinedAnnotations}
\alias{undefinedAnnotations}
\title{(Admin) Finds undefined REDCap variables}
\usage{
undefinedAnnotations()
}
\value{
A data frame that is a template to add to the commonKnowledge data in the tgr-annotations repo with new definitions or corrections.
}
\description{
Pulls sample data down from REDCap and compares them with the defined annotation list in GitHub and identifies variables in REDCap that need defining in GitHub.  Only variables for which there is a value in at least one record are returned.
}
\details{
Requires **admin** REDCap credentials to be set in the environment.
}
\author{
Amy Paguirigan
}
