% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgapi.R
\name{validate_date}
\alias{validate_date}
\title{Validate and format date inputs}
\usage{
validate_date(date)
}
\arguments{
\item{date}{Date input (string or Date object)}
}
\value{
Formatted date string
}
\description{
Validate and format date inputs
}
