% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgapi.R
\name{validate_fields}
\alias{validate_fields}
\title{Validate selected fields against schema}
\usage{
validate_fields(fields)
}
\arguments{
\item{fields}{Vector of field names}
}
\value{
Validated field names
}
\description{
Validate selected fields against schema
}
