% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgapi.R
\name{validate_prices}
\alias{validate_prices}
\title{Validate price parameters}
\usage{
validate_prices(min_price, max_price)
}
\arguments{
\item{min_price}{Minimum price}

\item{max_price}{Maximum price}
}
\value{
List of validated prices
}
\description{
Validate price parameters
}
