% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateDataset.R
\name{validateDataFrame}
\alias{validateDataFrame}
\title{Validate an existing data frame prior to upload}
\usage{
validateDataFrame(data, validation_id = NULL)
}
\arguments{
\item{data}{The data frame containing the data you want to validate.}

\item{validation_id}{The name of the id you want to validate the dataset with respect to:  subject_id, biospecimen_id or assay_material.}
}
\value{
A data frame containing any rows of the original data frame for which there are one or more mismatches between records with matching validation_ids.
}
\description{
Validate an existing data frame prior to upload
}
\details{
Requires setCreds() to have been run successfully.
}
\author{
Amy Paguirigan
}
