% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateDataset.R
\name{validateDataset}
\alias{validateDataset}
\title{Validate patient, specimen or assay material level datasets}
\usage{
validateDataset(validation_id = NULL)
}
\arguments{
\item{validation_id}{The name of the id you want to validate the dataset with respect to:  subject_id, biospecimen_id or assay_material.}
}
\value{
A data frame containing all of the records for which there are one or more mismatches between records with matching validation_ids.
}
\description{
Validate patient, specimen or assay material level datasets
}
\details{
Requires setCreds() to have been run successfully.
}
\author{
Amy Paguirigan
}
