% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepForUpload.R
\name{validateUpload}
\alias{validateUpload}
\title{Validate a data frame for upload to TGR}
\usage{
validateUpload(dataToValidate)
}
\arguments{
\item{dataToValidate}{A data frame containing the data you want to validate for upload.
Note:  For a valid dataset to be uploaded to the TGR REDCap, it must have the `molecular_id` column first.}
}
\value{
A list of results including: `assessment` which is the results from REDCapR's `validate_for_write` function;
`invalid` which is a data frame including `molecular_id` that includes columns that are not included in the TGR data dictionary and cannot be uploaded;
and `valid` which is a data frame suitable for uploading via the Data Import Tool in REDCap after writing it to a csv.
}
\description{
Validate a data frame for upload to TGR
}
\details{
Requires REDCap credentials to be set in the environment.  Does not allow for changes in DAG or marking if an instrument is complete.
}
\author{
Amy Paguirigan
}
