% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{annotationDictionary}
\alias{annotationDictionary}
\title{Create a snapshot of the Annotation Dictionary (not needed anymore)}
\usage{
annotationDictionary(commonKnowledge)
}
\arguments{
\item{commonKnowledge}{The commonKnowledge data frame containing current annotations via pullAnnotations().}
}
\value{
A character vector containing the `categorical` annotations in the TGR.
}
\description{
Pulls sample data down from REDCap in order to generate example column lists for use in Shiny UI's.
}
\details{
Requires **admin** REDCap credentials to be set in the environment.
}
\author{
Amy Paguirigan
}
