% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cromwellfunctions.R
\name{cromwellJobs}
\alias{cromwellJobs}
\title{Pull Cromwell Workflow Jobs}
\usage{
cromwellJobs(days = 1)
}
\arguments{
\item{days}{The number of days of history to return, defaults to 1 day.}
}
\value{
Returns a long form data frame of metadata on workflow jobs submitted
}
\description{
Requests metadata about Cromwell Workflow Jobs during a time period specified.
}
\details{
Requires valid Cromwell URL to be set in the environment.
}
\author{
Amy Paguirigan
}
