% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDictionary.R
\name{getDictionary}
\alias{getDictionary}
\title{Retrieve an example data dictionary from the TGR REDCap projects}
\usage{
getDictionary(project)
}
\arguments{
\item{project}{A string for which REDCap project to return data dictionary for, such as "specimen", "assay", or "molecular".}
}
\value{
An array of field names for the project specified.
}
\description{
Retrieve an example data dictionary from the TGR REDCap projects
}
\details{
Requires REDCap credentials to be set in the environment.
}
\author{
Amy Paguirigan
}
