% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listS3Objects.R
\name{listS3Objects}
\alias{listS3Objects}
\title{Pull the list of objects and tags in S3}
\usage{
listS3Objects(bucket)
}
\arguments{
\item{bucket}{The name of the S3 bucket containing the data.}
}
\value{
Returns a long form data frame of annotated objects in the S3 bucket.
}
\description{
Pulls Information from the results of s3tagcrawler for TGR that are in an S3 bucket, including the object list and their tags as well as size metadata.
}
\details{
Requires S3 credentials to be set in the environment by setCreds.
}
\author{
Amy Paguirigan
}
