% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgapi.R
\name{process_response}
\alias{process_response}
\title{Process JSON response into tibble}
\usage{
process_response(
  json,
  selected_fields = NULL,
  tmk = NULL,
  min_price = NULL,
  max_price = NULL
)
}
\arguments{
\item{json}{JSON response}

\item{selected_fields}{Fields to include in output}

\item{tmk}{Tax map key}

\item{min_price}{Minimum sales price}

\item{max_price}{Maximum sales price}
}
\value{
Tibble with selected fields
}
\description{
Process JSON response into tibble
}
