% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setCreds.R
\name{setCreds}
\alias{setCreds}
\title{Sets tgR environment variables}
\usage{
setCreds(tokenSet = "file", path)
}
\arguments{
\item{tokenSet}{Which token set to apply, default is "file", and then a path to a credentials file is required.  Other options include "admin" or "app".}

\item{path}{The path to the file to source that contains appropriately defined environment variables.}
}
\value{
Nothing, sets environment variables for the remaining functions to work.
}
\description{
Sets tgR environment variables
}
\details{
Requires the desired REDCap and S3 credentials to be set in the environment or be available in an R file specified.
}
\author{
Amy Paguirigan
}
