% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgapi.R
\name{tg}
\alias{tg}
\title{GET data from TG API}
\usage{
tg(startDate, endDate)
}
\arguments{
\item{startDate}{A date string of the format YYYY-MM-DD}

\item{endDate}{A date string of the format YYYY-MM-DD}
}
\value{
A tibble of data retrieved from the API
}
\description{
GET data from TG API
}
