% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgrAnnotate.R
\name{tgrAnnotate}
\alias{tgrAnnotate}
\title{Queries TGR REDCap for dataset annotations}
\usage{
tgrAnnotate(DAG = NULL, harmonizedOnly = FALSE, evenEmptyCols = FALSE)
}
\arguments{
\item{DAG}{(admin only) A character vector contining the name(s) of the TGR data access group(s) for which to request data, or "all".}

\item{harmonizedOnly}{(Optional) Whether you want only the annotations that are harmonized across all projects (TRUE) or a complete data set (FALSE, default).}

\item{evenEmptyCols}{(Optional) Whether you want even the empty fields in your final data frame (TRUE) or if you just want columns where there is at least one value in the resulting dataset (FALSE, default).}
}
\value{
Returns a long form data frame containing the requested dataset.
}
\description{
Queries TGR REDCap for dataset annotations
}
\details{
Requires setCreds() to have been run successfully.
}
\author{
Amy Paguirigan
}
