% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{undefinedAnnotations}
\alias{undefinedAnnotations}
\title{Finds undefined REDCap variables}
\usage{
undefinedAnnotations(commonKnowledge)
}
\arguments{
\item{commonKnowledge}{The commonKnowledge data frame containing current annotations via pullAnnotations().}
}
\value{
A data frame of makeMeaning.  This is a template to add to the commonKnowledge dataframe with new definitions or corrections.
}
\description{
Pulls sample data down from REDCap and compares them with the defined annotation list in GitHub and identifies variables in REDCap that need defining in GitHub.  Only variables for which there is a value in at least one record are returned.
}
\details{
Requires REDCap credentials to be set in the environment.
}
\author{
Amy Paguirigan
}
