% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{usedIdentifiers}
\alias{usedIdentifiers}
\title{Test if identifiers have already been used in the TGR.}
\usage{
usedIdentifiers(x, type)
}
\arguments{
\item{x}{Character vector of proposed identifiers.}

\item{type}{Identifier type: `biospecimen_id`, `assay_material_id`, or `molecular_id`.}
}
\value{
Returns the list of idnetifiers that have already been used.
}
\description{
Test if identifiers have already been used in the TGR.
}
\details{
Requires **admin** REDCap credentials to be set in the environment.
}
\author{
Amy Paguirigan
}
